+++ concordium-consensus/src/Concordium/KonsensusV1/Consensus/Finality.hs
@@ -21,6 +21,7 @@ import Concordium.Types.Parameters hiding (getChainParameters)
 import Concordium.Types.SeedState
 import Concordium.Utils
 
+import Concordium.Genesis.Data.BaseV1
 import Concordium.GlobalState.BlockState
 import qualified Concordium.GlobalState.Persistent.BlockState as PBS
 import Concordium.GlobalState.Statistics
@@ -89,13 +90,83 @@ processCertifiedBlock cb@CertifiedBlock{..}
         alreadyStored <- LowLevel.memberBlock (getHash cbQuorumBlock)
         unless alreadyStored a
 
+-- | Result of catching up a 'FinalizationEntry'.
+data CatchupFinalizationEntryResult
+    = -- | The 'FinalizationEntry' is valid and consensus
+      --  processed it.
+      CFERSuccess
+    | -- | The proposed finalized block is inconsistent
+      --  with the finalized block being pointed to in
+      --  the 'FinalizationEntry'.
+      CFERInconsistent
+    | -- | The finalization entry is invalid.
+      CFERInvalid
+    | -- | The finalization entry pointed to a block that was not
+      --  alive or the block is already finalized.
+      CFERNotAlive
+    | -- | The baking committee could not be looked up for the 'Epoch'
+      --  of the finalized 'QuorumCertificate'.
+      CFERUnknownBakers
+
+-- | Receive a 'FinalizationEntry' as part of catch-up.
+--  This function checks whether the finalization entry is
+--  consistent with the block indicated by @feFinalizedQuorumCertificate@ of the
+--  finalization entry and that block is alive and non finalized.
+--  If the finalization entry can be verified then it is processed.
+catchupFinalizationEntry ::
+    ( MonadState (SkovData (MPV m)) m,
+      TimeMonad m,
+      MonadIO m,
+      LowLevel.MonadTreeStateStore m,
+      BlockStateStorage m,
+      GSTypes.BlockState m ~ PBS.HashedPersistentBlockState (MPV m),
+      MonadThrow m,
+      MonadConsensusEvent m,
+      MonadLogger m,
+      IsConsensusV1 (MPV m)
+    ) =>
+    -- | The finalization entry received.
+    FinalizationEntry ->
+    -- | Whether the finalization entry was successfully processed.
+    m CatchupFinalizationEntryResult
+catchupFinalizationEntry finEntry = do
+    let finQC = feFinalizedQuorumCertificate finEntry
+    let finBlockHash = qcBlock finQC
+    gets (getLiveBlock finBlockHash) >>= \case
+        Just block
+            | blockRound block /= qcRound finQC || blockEpoch block /= qcEpoch finQC -> return CFERInconsistent
+            | otherwise ->
+                gets (getBakersForEpoch (qcEpoch finQC)) >>= \case
+                    Nothing -> return CFERUnknownBakers
+                    Just BakersAndFinalizers{..} -> do
+                        gm <- use genesisMetadata
+                        checkFinEntry _bfFinalizers gm $ do
+                            -- Record the checked quorum certificates
+                            recordCheckedQuorumCertificate $
+                                feFinalizedQuorumCertificate finEntry
+                            recordCheckedQuorumCertificate $
+                                feSuccessorQuorumCertificate finEntry
+                            -- Process the finalization
+                            processFinalizationEntry block finEntry
+                            return CFERSuccess
+        Nothing -> return CFERNotAlive
+  where
+    checkFinEntry finCommittee GenesisMetadata{..} cont =
+        if checkFinalizationEntry
+            gmCurrentGenesisHash
+            (toRational $ genesisSignatureThreshold gmParameters)
+            finCommittee
+            finEntry
+            then cont
+            else return CFERInvalid
+
 -- | Process a finalization entry that finalizes a block that is not currently considered finalized.
 --
---  PRECONDITION:
---   * The block is live and not already finalized.
---   * The finalization entry is valid.
---   * The block is at most one epoch later than the last finalized block. (This is implied by
---     the block being live.)
+--   PRECONDITION:
+--    * The block is live and not already finalized.
+--    * The finalization entry is valid.
+--    * The block is at most one epoch later than the last finalized block. (This is implied by
+--      the block being live.)
 processFinalizationEntry ::
     ( MonadState (SkovData (MPV m)) m,
       TimeMonad m,