+++ concordium-consensus/test-runners/catchup/Main.hs
@@ -241,8 +241,8 @@ peerReceive target src MessageBlock genIndex msg = do
     res <- runMVR (receiveExecuteBlock genIndex msg) (peerMVR target)
     when (isPending res) $ markPeerPending target (peerId src)
 peerReceive target src MessageFinalizationRecord genIndex msg = do
-    mvLog (peerMVR target) External LLDebug $ "Received finalization record from " ++ show (peerId src) ++ " genesisIndex:" ++ show genIndex
-    res <- runMVR (receiveFinalizationRecord genIndex msg) (peerMVR target)
+    mvLog (peerMVR target) External LLDebug $ "Received finalization entity from " ++ show (peerId src) ++ " genesisIndex:" ++ show genIndex
+    res <- runMVR (receiveFinalization genIndex msg) (peerMVR target)
     when (isPending res) $ markPeerPending target (peerId src)
 peerReceive target src MessageFinalization genIndex msg = do
     mvLog (peerMVR target) External LLDebug $ "Received finalization message from " ++ show (peerId src) ++ " genesisIndex:" ++ show genIndex